+++ CHANGELOG.md
@@ -1,3 +1,94 @@
+## [1.55.8](https://github.com/adobe/spacecat-audit-worker/compare/v1.55.7...v1.55.8) (2025-02-21)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-utils to v1.33.0 ([#684](https://github.com/adobe/spacecat-audit-worker/issues/684)) ([57cb5b9](https://github.com/adobe/spacecat-audit-worker/commit/57cb5b90831c316ad3cafdacf716c03e81900907))
+
+## [1.55.7](https://github.com/adobe/spacecat-audit-worker/compare/v1.55.6...v1.55.7) (2025-02-21)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-data-access to v2.8.2 ([#681](https://github.com/adobe/spacecat-audit-worker/issues/681)) ([016c3b9](https://github.com/adobe/spacecat-audit-worker/commit/016c3b9e41a513eba7605140912dd0af16bda4df))
+
+## [1.55.6](https://github.com/adobe/spacecat-audit-worker/compare/v1.55.5...v1.55.6) (2025-02-21)
+
+
+### Bug Fixes
+
+* sites-28998 fix for ignored entries are overwritten with next audit run ([#653](https://github.com/adobe/spacecat-audit-worker/issues/653)) ([0cfb009](https://github.com/adobe/spacecat-audit-worker/commit/0cfb00926c64af650cc20a19d2e2c3fd2565345e))
+
+## [1.55.5](https://github.com/adobe/spacecat-audit-worker/compare/v1.55.4...v1.55.5) (2025-02-20)
+
+
+### Bug Fixes
+
+* opportunity schema violation ([#679](https://github.com/adobe/spacecat-audit-worker/issues/679)) ([447f5b3](https://github.com/adobe/spacecat-audit-worker/commit/447f5b3868b89ad152beeb8c7b43104245a5c5c6))
+
+## [1.55.4](https://github.com/adobe/spacecat-audit-worker/compare/v1.55.3...v1.55.4) (2025-02-20)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-data-access to v2.8.1 ([#677](https://github.com/adobe/spacecat-audit-worker/issues/677)) ([2110793](https://github.com/adobe/spacecat-audit-worker/commit/21107930e4d96f83f3f2ff47f1b4f0ebec87a7ed))
+
+## [1.55.3](https://github.com/adobe/spacecat-audit-worker/compare/v1.55.2...v1.55.3) (2025-02-20)
+
+
+### Bug Fixes
+
+* chain step response ([#678](https://github.com/adobe/spacecat-audit-worker/issues/678)) ([62c30f1](https://github.com/adobe/spacecat-audit-worker/commit/62c30f12d3b0c9b15589da93cc224a7e432ea4d8))
+
+## [1.55.2](https://github.com/adobe/spacecat-audit-worker/compare/v1.55.1...v1.55.2) (2025-02-20)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-data-access to v2.8.0 ([#668](https://github.com/adobe/spacecat-audit-worker/issues/668)) ([1fbe759](https://github.com/adobe/spacecat-audit-worker/commit/1fbe759316acb169ec65e4549103998a3c3ab7ff))
+
+## [1.55.1](https://github.com/adobe/spacecat-audit-worker/compare/v1.55.0...v1.55.1) (2025-02-19)
+
+
+### Bug Fixes
+
+* guidance high organic low ctr ([#620](https://github.com/adobe/spacecat-audit-worker/issues/620)) ([3ae0969](https://github.com/adobe/spacecat-audit-worker/commit/3ae096959f65b734ca6fc6adea71daf119dce964))
+
+# [1.55.0](https://github.com/adobe/spacecat-audit-worker/compare/v1.54.1...v1.55.0) (2025-02-19)
+
+
+### Features
+
+* add image alt-text audit ([#647](https://github.com/adobe/spacecat-audit-worker/issues/647)) ([3e9309a](https://github.com/adobe/spacecat-audit-worker/commit/3e9309a4c7550a8454624f163b68724d8a4bed97))
+
+## [1.54.1](https://github.com/adobe/spacecat-audit-worker/compare/v1.54.0...v1.54.1) (2025-02-19)
+
+
+### Bug Fixes
+
+* step audit queue url ([#672](https://github.com/adobe/spacecat-audit-worker/issues/672)) ([15a0c33](https://github.com/adobe/spacecat-audit-worker/commit/15a0c339d60434154f89a3dfc8c9b4d61541462e))
+
+# [1.54.0](https://github.com/adobe/spacecat-audit-worker/compare/v1.53.2...v1.54.0) (2025-02-18)
+
+
+### Features
+
+* audit steps ([#659](https://github.com/adobe/spacecat-audit-worker/issues/659)) ([8b0362d](https://github.com/adobe/spacecat-audit-worker/commit/8b0362dae4c57440d9f786d4dfe74105df2124ca))
+
+## [1.53.2](https://github.com/adobe/spacecat-audit-worker/compare/v1.53.1...v1.53.2) (2025-02-18)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-data-access to v2.6.0 ([#667](https://github.com/adobe/spacecat-audit-worker/issues/667)) ([4fbd3db](https://github.com/adobe/spacecat-audit-worker/commit/4fbd3db2481c849e175873737aeef2a019c9ee6f))
+
+## [1.53.1](https://github.com/adobe/spacecat-audit-worker/compare/v1.53.0...v1.53.1) (2025-02-18)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#666](https://github.com/adobe/spacecat-audit-worker/issues/666)) ([10204dd](https://github.com/adobe/spacecat-audit-worker/commit/10204dd133f339aad0ea00645cfdbd1267c989ed))
+
 # [1.53.0](https://github.com/adobe/spacecat-audit-worker/compare/v1.52.15...v1.53.0) (2025-02-17)